Order spec

+++ migrations/sqls/20220109070357-orders-table-up.sql
@@ -2,5 +2,5 @@
 CREATE TABLE orders (
 	id SERIAL PRIMARY KEY,
 	order_status VARCHAR(30) NULL,
-  store_user_id INTEGER REFERENCES storeusers (id)
+  store_user_id bigint REFERENCES storeusers(id)
 );
